#134 [feat] 데이터 업로드 서비스 메서드를 구현한다.

+++ src/main/java/com/dataracy/modules/data/domain/model/Data.java
@@ -26,8 +26,8 @@ public class Data {
     private String analysisGuide;
     private String dataFileUrl;
     private String thumbnailUrl;
-    private Long downloadCount;
-    private Long recentWeekDownloadCount;
+    private int downloadCount;
+    private int recentWeekDownloadCount;
     private DataMetadataEntity metadata;
 
     public static Data toDomain(
@@ -43,8 +43,8 @@ public class Data {
             String analysisGuide,
             String dataFileUrl,
             String thumbnailUrl,
-            Long downloadCount,
-            Long recentWeekDownloadCount,
+            int downloadCount,
+            int recentWeekDownloadCount,
             DataMetadataEntity metadata
     ) {
         return Data.builder()
